Rename sync_test_api → test_api, assert_green_bang → assert_green
