+++ migration-engine/migration-engine-tests/src/lib.rs
@@ -1,7 +1,7 @@
 #![deny(rust_2018_idioms, unsafe_code)]
 
 pub mod multi_engine_test_api;
-pub mod sync_test_api;
+pub mod test_api;
 
 mod assertions;
 mod commands;
