+++ migration-engine/migration-engine-tests/tests/errors/error_tests.rs
@@ -1,6 +1,6 @@
 use indoc::{formatdoc, indoc};
 use migration_core::rpc_api;
-use migration_engine_tests::sync_test_api::*;
+use migration_engine_tests::test_api::*;
 use pretty_assertions::assert_eq;
 use quaint::prelude::Insert;
 use serde_json::json;
@@ -283,7 +283,7 @@ fn unique_constraint_errors_in_migrations_must_return_a_known_error(api: TestApi
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
 
     let insert = Insert::multi_into(api.render_table_name("Fruit"), &["name"])
         .values(("banana",))
