+++ migration-engine/migration-engine-tests/tests/evaluate_data_loss/evaluate_data_loss_tests.rs
@@ -1,5 +1,5 @@
 use migration_core::commands::EvaluateDataLossOutput;
-use migration_engine_tests::sync_test_api::*;
+use migration_engine_tests::test_api::*;
 use pretty_assertions::assert_eq;
 
 #[test_connector]
