+++ migration-engine/migration-engine-tests/tests/existing_data/mod.rs
@@ -2,7 +2,7 @@ mod sql_unexecutable_migrations;
 mod sqlite_existing_data_tests;
 mod type_migration_tests;
 
-use migration_engine_tests::sync_test_api::*;
+use migration_engine_tests::test_api::*;
 use pretty_assertions::assert_eq;
 use prisma_value::PrismaValue;
 use quaint::ast::*;
@@ -16,7 +16,7 @@ fn dropping_a_table_with_rows_should_warn(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
 
     let original_database_schema = api.assert_schema().into_schema();
 
@@ -47,7 +47,7 @@ fn dropping_a_column_with_non_null_values_should_warn(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
 
     let original_database_schema = api.assert_schema().into_schema();
 
@@ -85,7 +85,7 @@ fn altering_a_column_without_non_null_values_should_not_warn(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
 
     let original_database_schema = api.assert_schema().into_schema();
 
@@ -116,7 +116,7 @@ fn altering_a_column_with_non_null_values_should_warn(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
     let original_database_schema = api.assert_schema().into_schema();
 
     let insert = Insert::multi_into(api.render_table_name("Test"), vec!["id", "age"])
@@ -254,7 +254,7 @@ fn column_defaults_can_safely_be_changed(api: TestApi) {
                     .unwrap_or_else(String::new)
             );
 
-            api.schema_push_w_datasource(dm2).send().assert_green_bang();
+            api.schema_push_w_datasource(dm2).send().assert_green();
         }
 
         // Check that the data is still there
@@ -297,7 +297,7 @@ fn changing_a_column_from_required_to_optional_should_work(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
     let original_database_schema = api.assert_schema().into_schema();
 
     let insert = Insert::multi_into(api.render_table_name("Test"), &["id", "age"])
@@ -313,7 +313,7 @@ fn changing_a_column_from_required_to_optional_should_work(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm2).send().assert_green_bang();
+    api.schema_push_w_datasource(dm2).send().assert_green();
 
     api.assert_schema().assert_ne(&original_database_schema);
 
@@ -339,7 +339,7 @@ fn changing_a_column_from_optional_to_required_is_unexecutable(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm).send().assert_green_bang();
+    api.schema_push_w_datasource(dm).send().assert_green();
     let original_database_schema = api.assert_schema().into_schema();
 
     let insert = Insert::multi_into(api.render_table_name("Test"), &["id", "age"])
@@ -398,7 +398,7 @@ fn dropping_a_table_referenced_by_foreign_keys_must_work(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     api.assert_schema()
         .assert_table("Category", |table| table.assert_columns_count(2))
@@ -444,7 +444,7 @@ fn string_columns_do_not_get_arbitrarily_migrated(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     let insert = Insert::single_into(api.render_table_name("User"))
         .value("id", "the-id")
@@ -464,10 +464,7 @@ fn string_columns_do_not_get_arbitrarily_migrated(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm2)
-        .send()
-        .assert_green_bang()
-        .assert_green_bang();
+    api.schema_push_w_datasource(dm2).send().assert_green().assert_green();
 
     // Check that the string values are still there.
     let select = Select::from_table(api.render_table_name("User"))
@@ -497,7 +494,7 @@ fn altering_the_type_of_a_column_in_an_empty_table_should_not_warn(api: TestApi)
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     let dm2 = r#"
         model User {
@@ -507,7 +504,7 @@ fn altering_the_type_of_a_column_in_an_empty_table_should_not_warn(api: TestApi)
         }
     "#;
 
-    api.schema_push_w_datasource(dm2).send().assert_green_bang();
+    api.schema_push_w_datasource(dm2).send().assert_green();
 
     api.assert_schema().assert_table("User", |table| {
         table.assert_column("dogs", |col| col.assert_type_is_string().assert_is_required())
@@ -524,7 +521,7 @@ fn making_a_column_required_in_an_empty_table_should_not_warn(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     let dm2 = r#"
         model User {
@@ -534,7 +531,7 @@ fn making_a_column_required_in_an_empty_table_should_not_warn(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm2).send().assert_green_bang();
+    api.schema_push_w_datasource(dm2).send().assert_green();
 
     api.assert_schema().assert_table("User", |table| {
         table.assert_column("dogs", |col| col.assert_type_is_int().assert_is_required())
@@ -563,7 +560,7 @@ fn enum_variants_can_be_added_without_data_loss(api: TestApi) {
     api.schema_push_w_datasource(dm1)
         .migration_id(Some("initial-setup"))
         .send()
-        .assert_green_bang();
+        .assert_green();
 
     {
         let cat_inserts = quaint::ast::Insert::multi_into(api.render_table_name("Cat"), vec!["id", "mood"])
@@ -594,7 +591,7 @@ fn enum_variants_can_be_added_without_data_loss(api: TestApi) {
     api.schema_push_w_datasource(dm2)
         .migration_id(Some("add-absolutely-fabulous-variant"))
         .send()
-        .assert_green_bang();
+        .assert_green();
 
     // Assertions
     {
@@ -667,7 +664,7 @@ fn enum_variants_can_be_dropped_without_data_loss(api: TestApi) {
     api.schema_push_w_datasource(dm1)
         .migration_id(Some("initial-setup"))
         .send()
-        .assert_green_bang();
+        .assert_green();
 
     {
         let cat_inserts = quaint::ast::Insert::multi_into(api.render_table_name("Cat"), &["id", "mood"])
@@ -762,7 +759,7 @@ fn set_default_current_timestamp_on_existing_column_works(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     let insert = Insert::single_into(api.render_table_name("User")).value("id", 5).value(
         "created_at",
@@ -809,7 +806,7 @@ fn primary_key_migrations_do_not_cause_data_loss(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     api.insert("Dog")
         .value("name", "Marnie")
@@ -888,7 +885,7 @@ fn failing_enum_migrations_should_not_be_partially_applied(api: TestApi) {
     api.schema_push_w_datasource(dm1)
         .migration_id(Some("initial-setup"))
         .send()
-        .assert_green_bang();
+        .assert_green();
 
     {
         let cat_inserts = quaint::ast::Insert::multi_into(api.render_table_name("Cat"), &["id", "mood"])
