+++ migration-engine/migration-engine-tests/tests/existing_data/sql_unexecutable_migrations/made_optional_field_required.rs
@@ -1,4 +1,4 @@
-use migration_engine_tests::sync_test_api::*;
+use migration_engine_tests::test_api::*;
 use quaint::Value;
 use sql_schema_describer::DefaultValue;
 
@@ -12,7 +12,7 @@ fn making_an_optional_field_required_with_data_without_a_default_is_unexecutable
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     api.insert("Test")
         .value("id", "abc")
@@ -54,7 +54,7 @@ fn making_an_optional_field_required_with_data_with_a_default_works(api: TestApi
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     api.insert("Test")
         .value("id", "abc")
@@ -111,7 +111,7 @@ fn making_an_optional_field_required_with_data_with_a_default_is_unexecutable(ap
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     let initial_schema = api.assert_schema().into_schema();
 
@@ -170,7 +170,7 @@ fn making_an_optional_field_required_on_an_empty_table_works(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm1).send().assert_green_bang();
+    api.schema_push_w_datasource(dm1).send().assert_green();
 
     let dm2 = r#"
         model Test {
@@ -180,7 +180,7 @@ fn making_an_optional_field_required_on_an_empty_table_works(api: TestApi) {
         }
     "#;
 
-    api.schema_push_w_datasource(dm2).send().assert_green_bang();
+    api.schema_push_w_datasource(dm2).send().assert_green();
 
     api.assert_schema()
         .assert_table("Test", |table| table.assert_does_not_have_column("Int"));
